Initial Spike
https://trello.com/c/M6hL2wKC

This commit adds an initial implementation that was tested by hand.

Once we get a better sense of how Linters are expected to behave, we can
throw away the spike and test drive the new implementation.

To test locally:

```bash
$ cd node_modules/node-resque/resque-web
$ bundle
$ bundle exec rackup
$ open http://localhost:9292
```

Then, kill `foreman`, uncomment the block at the bottom of `index.js`,
and `foreman start`. Monitor the Resque queue for a
`"JavaScriptReviewJob"` in the `"java_script_review"` queue, as well as
an outgoing `"CompletedFileReviewJob"` in the `"high"` queue.